refactor: 명확한 네이밍으로 변경 및 final 키워드 추가 및 util 패키지 생성
- LottoParser가 domain/lotto에 포함되어있었는데 다른 곳에서도 사용되는 것을 고려해 패키지를 분리하였습니다.
