+++ src/main/java/view/InputView.java
@@ -22,7 +22,7 @@ public final class InputView {
         return scanner.nextLine();
     }
 
-    public static List<String> inputManualLottoNumbers(int count) {
+    public static List<String> inputManualLottoNumbers(final int count) {
         System.out.println();
         System.out.println("수동으로 구매할 번호를 입력해 주세요.");
         List<String> manualLottos = new ArrayList<>();
